split some util tests into their own files.
